Revert "Merge branch 'song-share-mvp' into FM-115_implement-song-sharing-backend"
This reverts commit 753bca289d13394cfdcd04c51d713fa3d8814149, reversing
changes made to 5b4af3b9ec77cca9812e5fee876ffb7915d3b2cd.
