+++ .github/pull_request_template.md
@@ -1,9 +1,2 @@
-### Checklist
-- [ ] Added documentation to all classes and functions
-
-**If merging to `main`**
-  - [ ] Incremented the version number
-  - [ ] Created a new changelog view
-
-### Changes
+#### Changes
 -
