+++ spoti-friends/src/common/views/PageTitle.swift
@@ -13,6 +13,7 @@ struct PageTitle: View {
             
             Spacer()
         }
+        .padding(.leading, 8)
     }
 }
