--- spoti-friends/src/common/views/settings/changelogs/v1.3.1.swift
@@ -1,25 +1,0 @@
-import SwiftUI
-
-struct v1_3_1: View {
-    var body: some View {
-        VStack (alignment: .leading) {
-            HStack {
-                Text("[v1.3.1]")
-                    .font(.title2)
-                    .bold()
-                Spacer()
-            }
-            .padding(.bottom, 4)
-            Text(
-"""
-• Prepared app for App Store release
-• Added a settings page
-"""
-            )
-        }
-    }
-}
-
-#Preview {
-    v1_3_1()
-}
