+++ spoti-friends/src/profile/views/ProfileView.swift
@@ -28,7 +28,6 @@ struct ProfileView: View {
                 } else {
                     NonUserProfileView(profile: profile)
                         .environmentObject(profileViewModel)
-                        .padding(.horizontal, 20)
                 }
             }
             .frame(minHeight: reader.size.height)
