+++ spoti-friends/src/profile/views/viewMore/ViewMoreTopSongs.swift
@@ -21,7 +21,7 @@ struct ViewMoreTopSongs: View {
     
     var body: some View {
         ScrollView {
-            VStack(alignment: .center) {
+            VStack(alignment: .leading) {
                 Text("Top Songs")
                     .foregroundStyle(Color.PresetColour.whitePrimary)
                     .font(.title2)
@@ -34,8 +34,6 @@ struct ViewMoreTopSongs: View {
                 .padding(.vertical, 8)
                 
                 TrackList(tracks: topTracks.tracks, showItemNumbers: true)
-                OpenSpotifyButton()
-                    .padding(.vertical, 12)
             }
             .frame(maxWidth: .infinity, alignment: .leading)
             .padding(.horizontal, 20)
